Resolve "Implement VM Controls"

Closes #23

See merge request learntech-lufgi9/bama/webwriter/webwriter-block-based-code!22

+++ widgets/lib/vm/vm.ts
@@ -3,11 +3,13 @@ export abstract class VirtualMachine {
 
   private highlightCallback: (id: string) => void;
 
-  public start(code: string): void {
-    this.initWorker(code);
+  public start(code: string, delay: number): void {
+    this.initWorker(code, delay);
   }
 
   public stop(): void {
+    if (!this.worker) return;
+
     this.highlight(null);
     this.worker.terminate();
   }
@@ -23,8 +25,8 @@ export abstract class VirtualMachine {
     ];
   }
 
-  private initWorker(code: string): void {
-    const script = this.generateWorkerScript(code);
+  private initWorker(code: string, delay: number): void {
+    const script = this.generateWorkerScript(code, delay);
     const url = this.generateWorkerScriptUrl(script);
     this.worker = new Worker(url);
     this.worker.onmessage = (event: MessageEvent<{ type: string, args: any[] }>) => {
@@ -35,27 +37,13 @@ export abstract class VirtualMachine {
     };
   }
 
-  private generateWorkerScript(code: string): string {
+  private generateWorkerScript(code: string, delay: number): string {
     let script = "";
-    const scriptFunction = (function () {
-      let resultResolveFunction: (result: any) => void;
+    script += "let resultResolveFunction;\n";
+    script += "async function wait(s) { await new Promise((resolve) => { setTimeout(resolve, s * 1e3) }); }\n";
+    script += `async function delay() { await new Promise((resolve) => { setTimeout(resolve, ${delay === 0 ? 40 : delay}) }); }\n`;
+    script += "onmessage = function (event) { if (event.data.type === 'result') { resultResolveFunction(event.data.args[0]); } };\n";
 
-      async function wait(s: number): Promise<void> {
-        await new Promise((resolve) => { setTimeout(resolve, s * 1e3); });
-      }
-
-      async function delay(ms: number): Promise<void> {
-        await new Promise((resolve) => { setTimeout(resolve, ms); });
-      }
-
-      onmessage = function (event) {
-        console.log(event);
-        if (event.data.type === "result") {
-          resultResolveFunction(event.data.args[0]);
-        }
-      };
-    });
-    script += `${scriptFunction.toString().match(/function[^{]+\{([\s\S]*)\}$/)[1]}\n`;
     this.callables.forEach((callable) => {
       const args = Array(callable.length).fill("x").map((x, i) => `${x}${i}`).join(", ");
       const message = `{ type: "${callable.name}", args: [${args}] }`;
@@ -69,7 +57,6 @@ export abstract class VirtualMachine {
     script += code;
     script += "highlight(null);\n";
     script += "})()\n";
-    console.log(script);
     return script;
   }
